requirements.txt
Django==1.10.2
django-cors-headers==1.2.1
django-environ==0.4.1
django-friendship==1.5.0
djangorestframework==3.4.7
gunicorn==19.7.0
httpie==0.9.6
httpsig==1.1.2
json-logging-py==0.2
Pillow==3.4.2
psycopg2==2.7.1
pycrypto==2.6.1
Pygments==2.1.3
PyJWT==1.3.0
pytz==2016.10
readline==6.2.4.1
requests==2.11.1
six==1.10.0
spotipy==2.3.8
stream-django==1.3.1
stream-python==2.3.9
